% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB1copula.R
\docType{class}
\name{BB1Copula-class}
\alias{BB1Copula-class}
\alias{dduCopula,numeric,BB1Copula-method}
\alias{ddvCopula,numeric,BB1Copula-method}
\alias{dduCopula,matrix,BB1Copula-method}
\alias{ddvCopula,matrix,BB1Copula-method}
\alias{getKendallDistr,BB1Copula-method}
\alias{kendallDistribution,BB1Copula-method}
\alias{surBB1Copula-class}
\alias{dduCopula,numeric,surBB1Copula-method}
\alias{ddvCopula,numeric,surBB1Copula-method}
\alias{dduCopula,matrix,surBB1Copula-method}
\alias{ddvCopula,matrix,surBB1Copula-method}
\alias{r90BB1Copula-class}
\alias{dduCopula,numeric,r90BB1Copula-method}
\alias{ddvCopula,numeric,r90BB1Copula-method}
\alias{dduCopula,matrix,r90BB1Copula-method}
\alias{ddvCopula,matrix,r90BB1Copula-method}
\alias{r270BB1Copula-class}
\alias{dduCopula,numeric,r270BB1Copula-method}
\alias{ddvCopula,numeric,r270BB1Copula-method}
\alias{dduCopula,matrix,r270BB1Copula-method}
\alias{ddvCopula,matrix,r270BB1Copula-method}
\title{BB1 copula models}
\description{
Wrapper classes representing the BB1, survival BB1, 90 degree and 270 degree
rotated BB1 copula families (Joe 1997) from \code{\link[=VineCopula-package]{VineCopula-package()}}.
}
\section{Objects from the Classes}{
 Objects can be created by calls of the
form \code{new("BB1Copula", ...)}, \code{new("surBB1Copula", ...)},
\code{new("r90BB1Copula", ...)} and \code{new("r270BB1Copula", ...)} or by the
functions \code{\link[=BB1Copula]{BB1Copula()}}, \code{\link[=surBB1Copula]{surBB1Copula()}}, \code{\link[=r90BB1Copula]{r90BB1Copula()}} and
\code{\link[=r270BB1Copula]{r270BB1Copula()}}.
}

\examples{

showClass("BB1Copula")
}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \linkS4class{BB6Copula}, \linkS4class{BB7Copula}, \linkS4class{BB8Copula} and
\linkS4class{joeCopula} for further wrapper classes to the
\code{\link[=VineCopula-package]{VineCopula-package()}}.
}
\keyword{classes}
