% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB6copula.R
\docType{class}
\name{BB6Copula-class}
\alias{BB6Copula-class}
\alias{dduCopula,numeric,BB6Copula-method}
\alias{ddvCopula,numeric,BB6Copula-method}
\alias{dduCopula,matrix,BB6Copula-method}
\alias{ddvCopula,matrix,BB6Copula-method}
\alias{getKendallDistr,BB6Copula-method}
\alias{kendallDistribution,BB6Copula-method}
\alias{surBB6Copula-class}
\alias{dduCopula,numeric,surBB6Copula-method}
\alias{ddvCopula,numeric,surBB6Copula-method}
\alias{dduCopula,matrix,surBB6Copula-method}
\alias{ddvCopula,matrix,surBB6Copula-method}
\alias{r90BB6Copula-class}
\alias{dduCopula,numeric,r90BB6Copula-method}
\alias{ddvCopula,numeric,r90BB6Copula-method}
\alias{dduCopula,matrix,r90BB6Copula-method}
\alias{ddvCopula,matrix,r90BB6Copula-method}
\alias{r270BB6Copula-class}
\alias{dduCopula,numeric,r270BB6Copula-method}
\alias{ddvCopula,numeric,r270BB6Copula-method}
\alias{dduCopula,matrix,r270BB6Copula-method}
\alias{ddvCopula,matrix,r270BB6Copula-method}
\title{BB6 copula models}
\description{
Wrapper classes representing the BB6, survival BB6, 90 degree and 270 degree
rotated BB6 copula families (Joe 1997) from \code{\link[=VineCopula-package]{VineCopula-package()}}.
}
\section{Objects from the Classes}{
 Objects can be created by calls of the
form \code{new("BB6Copula", ...)}, \code{new("surBB6Copula", ...)},
\code{new("r90BB6Copula", ...)} and \code{new("r270BB6Copula", ...)} or by the
functions \code{\link[=BB6Copula]{BB6Copula()}}, \code{\link[=surBB6Copula]{surBB6Copula()}}, \code{\link[=r90BB6Copula]{r90BB6Copula()}} and
\code{\link[=r270BB6Copula]{r270BB6Copula()}}.
}

\examples{

showClass("BB6Copula")
}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \linkS4class{BB6Copula}, \linkS4class{BB7Copula}, \linkS4class{BB8Copula} and
\linkS4class{joeCopula} for further wrapper classes to the
\code{\link[=VineCopula-package]{VineCopula-package()}}.
}
\keyword{classes}
