% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB8copula.R
\docType{class}
\name{BB8Copula-class}
\alias{BB8Copula-class}
\alias{dduCopula,numeric,BB8Copula-method}
\alias{ddvCopula,numeric,BB8Copula-method}
\alias{dduCopula,matrix,BB8Copula-method}
\alias{ddvCopula,matrix,BB8Copula-method}
\alias{getKendallDistr,BB8Copula-method}
\alias{kendallDistribution,BB8Copula-method}
\alias{surBB8Copula-class}
\alias{dduCopula,numeric,surBB8Copula-method}
\alias{ddvCopula,numeric,surBB8Copula-method}
\alias{dduCopula,matrix,surBB8Copula-method}
\alias{ddvCopula,matrix,surBB8Copula-method}
\alias{r90BB8Copula-class}
\alias{dduCopula,numeric,r90BB8Copula-method}
\alias{ddvCopula,numeric,r90BB8Copula-method}
\alias{dduCopula,matrix,r90BB8Copula-method}
\alias{ddvCopula,matrix,r90BB8Copula-method}
\alias{r270BB8Copula-class}
\alias{dduCopula,numeric,r270BB8Copula-method}
\alias{ddvCopula,numeric,r270BB8Copula-method}
\alias{dduCopula,matrix,r270BB8Copula-method}
\alias{ddvCopula,matrix,r270BB8Copula-method}
\title{BB8 copula models}
\description{
Wrapper classes representing the BB8, survival BB8, 90 degree and 270 degree
rotated BB8 copula families (Joe 1997) from \code{\link[=VineCopula-package]{VineCopula-package()}}.
}
\section{Objects from the Classes}{
 Objects can be created by calls of the
form \code{new("BB8Copula", ...)}, \code{new("surBB8Copula", ...)},
\code{new("r90BB8Copula", ...)} and \code{new("r270BB8Copula", ...)} or by the
functions \code{\link[=BB8Copula]{BB8Copula()}}, \code{\link[=surBB8Copula]{surBB8Copula()}}, \code{\link[=r90BB8Copula]{r90BB8Copula()}} and
\code{\link[=r270BB8Copula]{r270BB8Copula()}}.
}

\examples{

showClass("BB8Copula")
}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \linkS4class{BB8Copula}, \linkS4class{BB8Copula}, \linkS4class{BB8Copula} and
\linkS4class{joeCopula} for further wrapper classes to the
\code{\link[=VineCopula-package]{VineCopula-package()}}.
}
\keyword{classes}
