% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC2copula-package.R
\docType{package}
\name{VC2copula-package}
\alias{VC2copula}
\alias{VC2copula-package}
\title{VC2copula: Objects conversion between package VineCopula and copula}
\description{
Extends the 'copula' package with families and models from 
  'VineCopula'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tnagler/VC2copula}
  \item Report bugs at \url{https://github.com/tnagler/VC2copula/issues}
}

}
\author{
\strong{Maintainer}: Thomas Nagler \email{mail@tnagler.com}

}
\keyword{internal}
