% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC2copula-package.R
\docType{package}
\name{VC2copula-package}
\alias{VC2copula}
\alias{VC2copula-package}
\title{VC2copula: Extend the 'copula' Package with Families and Models from 'VineCopula'}
\description{
Provides new classes for (rotated) BB1, BB6, BB7, BB8, and Tawn copulas, extends the existing Gumbel and Clayton families with rotations, and allows to set up a vine copula model using the 'copula' API. Corresponding objects from the 'VineCopula' API can easily be converted.

Provides new classes for (rotated) BB1, BB6, BB7, BB8, and Tawn copulas, extends the existing Gumbel and Clayton families with rotations, and allows to set up a vine copula model using the 'copula' API. Corresponding objects from the 'VineCopula' API can easily be converted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tnagler/VC2copula}
  \item Report bugs at \url{https://github.com/tnagler/VC2copula/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/tnagler/VC2copula}
  \item Report bugs at \url{https://github.com/tnagler/VC2copula/issues}
}

}
\author{
\strong{Maintainer}: Thomas Nagler \email{mail@tnagler.com}

Other contributors:
\itemize{
  \item Benedikt Graeler [contributor]
}

}
\keyword{internal}
